add authentication app for cleaner project directory
