+++ graphify/settings.py
@@ -113,6 +113,7 @@ INSTALLED_APPS = [
     'django.contrib.messages',
     'django.contrib.staticfiles',
     'main',
+    'authentication',
     'corsheaders',
     'django_ratelimit',
 ]
