+++ main/urls.py
@@ -1,13 +1,10 @@
 from django.urls import path
-from main import views
+from django.views.generic import RedirectView
+from . import views
 
 urlpatterns = [
-    path('', views.login_page, name='login'),
+    # Automatically redirect the root URL to the login page
+    path('', RedirectView.as_view(pattern_name='login', permanent=False), name='home_redirect'),
+    
     path('dashboard/', views.dashboard_view, name='dashboard'),
-    path('register/', views.register_view, name='register'),
-    path('verify-otp/', views.otp_verify_view, name='verify_otp'),
-    path('auth/google/', views.google_login, name='google_login'), # JS hits this
-    path('auth/github/', views.github_login, name='github_login'),
-    # path('auth/github/callback/', views.github_callback, name='github_callback'),
-    path('logout/', views.logout_view, name='logout'),
 ]
