+++ main/views.py
@@ -1,440 +1,7 @@
-import json
-import logging
-import random
-import time
-import secrets
-import hashlib
-import requests
-import re
-from django.http import JsonResponse
-from django.shortcuts import render, redirect
-from django.views.decorators.csrf import csrf_exempt, csrf_protect
-from django.views.decorators.http import require_http_methods
-from django.views.decorators.cache import cache_page
-from django_ratelimit.decorators import ratelimit
-from firebase_admin import auth, firestore
-from django.contrib.auth.models import User
-from django.contrib.auth import login, logout
+from django.shortcuts import render
 from django.contrib.auth.decorators import login_required
-from django.conf import settings
-from django.urls import reverse
-from django.contrib import messages
-from django.core.mail import send_mail
-from django.core.exceptions import ValidationError
-from django.contrib.auth.password_validation import validate_password
 
-# Security Helpers
-def validate_email_format(email):
-    """Validates email format using RFC 5322 simplified pattern."""
-    email_pattern = r'^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$'
-    return re.match(email_pattern, email) is not None
-
-def validate_password_strength(password):
-    """Validates password meets minimum requirements."""
-    if len(password) < 8:
-        raise ValidationError("Password must be at least 8 characters long.")
-    if not any(char.isdigit() for char in password):
-        raise ValidationError("Password must contain at least one digit.")
-    if not any(char.isupper() for char in password):
-        raise ValidationError("Password must contain at least one uppercase letter.")
-    if not any(char in '!@#$%^&*()_+-=[]{}|;:,.<>?' for char in password):
-        raise ValidationError("Password must contain at least one special character.")
-
-@csrf_protect
-@ratelimit(key='ip', rate='10/h', method='POST')
-def login_page(request):
-    """Handles both displaying the login page and processing login attempts."""
-    if request.user.is_authenticated:
-        return redirect('dashboard')
-
-    if request.method == 'POST':
-        email = request.POST.get('email')
-        password = request.POST.get('password')
-        
-        if not email or not password:
-            messages.error(request, 'Email and password are required.')
-            return render(request, 'main/index.html', {
-                'firebase_config': json.dumps(settings.FIREBASE_PUBLIC_CONFIG)
-            })
-        
-        try:
-            # Verify user exists in Firebase Auth
-            firebase_user = auth.get_user_by_email(email)
-            
-            # Create or get Django user for session management
-            django_user, created = User.objects.get_or_create(
-                username=firebase_user.uid,
-                defaults={
-                    'email': email,
-                    'first_name': firebase_user.display_name.split(' ')[0] if firebase_user.display_name else '',
-                    'last_name': ' '.join(firebase_user.display_name.split(' ')[1:]) if firebase_user.display_name and ' ' in firebase_user.display_name else ''
-                }
-            )
-            
-            # Password verification should be done on client-side via Firebase SDK
-            # Here we just confirm the user exists in Firebase
-            # The actual password check happens client-side, and the user is logged in via Firebase
-            login(request, django_user)
-            return redirect('dashboard')
-            
-        except auth.UserNotFoundError:
-            messages.error(request, f'No account found for {email}. Please register.')
-            return redirect('register')
-        except Exception as e:
-            logging.error(f"Login error: {e}")
-            messages.error(request, 'Login failed. Please try again.')
-
-    # Securely pass the config to the template
-    context = {
-        'firebase_config': json.dumps(settings.FIREBASE_PUBLIC_CONFIG)
-    }
-    return render(request, 'main/index.html', context)
 @login_required
 def dashboard_view(request):
-    # The user object is automatically available in the template context
-    # when using RequestContext, which is default.
+    """Displays the main dashboard for logged-in users."""
     return render(request, 'main/dashboard.html')
-
-@csrf_exempt
-@ratelimit(key='ip', rate='20/h', method='POST')
-def google_login(request):
-    """Handles Google OAuth authentication via Firebase - SECURE VERSION."""
-    if request.method == "POST":
-        try:
-            data = json.loads(request.body)
-            id_token = data.get('id_token')
-
-            if not id_token:
-                return JsonResponse({"status": "error", "message": "ID token not provided."}, status=400)
-
-            # Verify the ID token against Firebase
-            decoded_token = auth.verify_id_token(id_token)
-            firebase_uid = decoded_token.get('uid')
-            email = decoded_token.get('email')
-            name = decoded_token.get('name', '')
-
-            if not email:
-                return JsonResponse({"status": "error", "message": "Email not provided by Google."}, status=400)
-
-            # Get or create Django user for session management (use Firebase UID as username)
-            django_user, created = User.objects.get_or_create(
-                username=firebase_uid,
-                defaults={
-                    'email': email,
-                    'first_name': name.split(' ')[0] if name else '',
-                    'last_name': ' '.join(name.split(' ')[1:]) if ' ' in name else ''
-                }
-            )
-
-            # Set unusable password for social auth users
-            if created or django_user.has_usable_password():
-                django_user.set_unusable_password()
-                django_user.save()
-
-            # Store user data in Firestore with Firebase UID as document ID
-            db = firestore.client()
-            user_doc = db.collection('users').document(firebase_uid)
-            user_doc.set({
-                'email': email,
-                'name': name,
-                'firebase_uid': firebase_uid,
-                'providers': firestore.ArrayUnion(['google']),
-                'last_login': firestore.SERVER_TIMESTAMP
-            }, merge=True)
-
-            # Log the user into Django's session framework
-            login(request, django_user)
-
-            return JsonResponse({"status": "success"})
-        except auth.InvalidIdTokenError:
-            logging.warning("Invalid ID token received (non-sensitive)")
-            return JsonResponse({"status": "error", "message": "Invalid authentication token."}, status=401)
-        except auth.ExpiredIdTokenError:
-            logging.warning("Expired ID token received (non-sensitive)")
-            return JsonResponse({"status": "error", "message": "Authentication token expired. Please try again."}, status=401)
-        except Exception as e:
-            # DO NOT log the full exception - it may contain sensitive data
-            logging.error(f"Google login error (non-sensitive): Authentication failed")
-            return JsonResponse({"status": "error", "message": "Authentication failed. Please try again."}, status=401)
-            
-    return JsonResponse({"status": "error", "message": "Method not allowed"}, status=405)
-
-def logout_view(request):
-    """Logs the user out and redirects to the login page."""
-    logout(request)
-    return redirect('login')
-
-@csrf_exempt
-@ratelimit(key='ip', rate='20/h', method='POST')
-def github_login(request):
-    """
-    Handles Firebase GitHub authentication token verification.
-    Firebase console manages GitHub OAuth directly.
-    """
-    if request.method == "POST":
-        try:
-            data = json.loads(request.body)
-            id_token = data.get('id_token')  # Firebase returns ID token
-
-            if not id_token:
-                return JsonResponse({"status": "error", "message": "ID token not provided."}, status=400)
-
-            # Verify the Firebase token
-            decoded_token = auth.verify_id_token(id_token)
-            firebase_uid = decoded_token.get('uid')
-            email = decoded_token.get('email')
-            name = decoded_token.get('name', '')
-
-            if not email:
-                return JsonResponse({"status": "error", "message": "Email not provided by GitHub."}, status=400)
-
-            # Get or create Django user for session management
-            django_user, created = User.objects.get_or_create(
-                username=firebase_uid,
-                defaults={
-                    'email': email,
-                    'first_name': name.split(' ')[0] if name else '',
-                    'last_name': ' '.join(name.split(' ')[1:]) if ' ' in name else ''
-                }
-            )
-
-            # Set unusable password for social auth users
-            if created or django_user.has_usable_password():
-                django_user.set_unusable_password()
-                django_user.save()
-
-            # Update Firestore with latest provider info
-            db = firestore.client()
-            user_doc = db.collection('users').document(firebase_uid)
-            user_doc.set({
-                'email': email,
-                'name': name,
-                'firebase_uid': firebase_uid,
-                'providers': firestore.ArrayUnion(['github']),
-                'last_login': firestore.SERVER_TIMESTAMP
-            }, merge=True)
-
-            # Log the user into Django
-            login(request, django_user)
-            return JsonResponse({"status": "success"})
-
-        except auth.InvalidIdTokenError:
-            logging.warning("Invalid GitHub ID token (non-sensitive)")
-            return JsonResponse({"status": "error", "message": "Invalid authentication token."}, status=401)
-        except auth.ExpiredIdTokenError:
-            logging.warning("Expired GitHub ID token (non-sensitive)")
-            return JsonResponse({"status": "error", "message": "Authentication token expired. Please try again."}, status=401)
-        except Exception as e:
-            logging.error(f"GitHub verification error (non-sensitive): {str(e)}")
-            return JsonResponse({"status": "error", "message": "Authentication failed. Please try again."}, status=401)
-
-    return JsonResponse({"status": "error", "message": "Method not allowed"}, status=405)
-@csrf_protect
-@ratelimit(key='ip', rate='5/h', method='POST')
-def register_view(request):
-    """Handles user registration and sending OTP - SECURE VERSION."""
-    if request.method == 'POST':
-        email = request.POST.get('email', '').strip()
-        password = request.POST.get('password', '')
-        password_conf = request.POST.get('password_conf', '')
-
-        # Validate email format
-        if not email or not validate_email_format(email):
-            messages.error(request, "Please enter a valid email address.")
-            return render(request, 'main/register.html')
-
-        # Validate password strength
-        if not password:
-            messages.error(request, "Password is required.")
-            return render(request, 'main/register.html')
-        
-        if password != password_conf:
-            messages.error(request, "Passwords do not match.")
-            return render(request, 'main/register.html')
-        
-        try:
-            validate_password_strength(password)
-        except ValidationError as e:
-            messages.error(request, str(e))
-            return render(request, 'main/register.html')
-
-        # Check if user already exists in Firebase
-        try:
-            existing_user = auth.get_user_by_email(email)
-            messages.error(request, "An account with this email already exists. Please login instead.")
-            return render(request, 'main/register.html')
-        except auth.UserNotFoundError:
-            pass  # Good, user doesn't exist
-        except Exception as e:
-            logging.error(f"Firebase lookup error (non-sensitive): User check failed")
-            messages.error(request, "Registration service temporarily unavailable. Please try again.")
-            return render(request, 'main/register.html')
-
-        # Generate OTP
-        otp = str(random.randint(100000, 999999))
-        secure_token = secrets.token_urlsafe(32)
-        otp_hash = hashlib.sha256(otp.encode()).hexdigest()
-        
-        # SECURITY: Store ONLY hashed data + temporary token reference
-        # NO plaintext password stored in session
-        request.session['registration_data'] = {
-            'email': email,
-            'otp_hash': otp_hash,
-            'secure_token': secure_token,
-            'created_at': time.time(),
-            'otp_expiry': time.time() + 300,  # 5-minute expiry
-        }
-        
-        # Store password hash (not plaintext) for verification later
-        # This allows us to verify password was correct without storing plaintext
-        password_hash = hashlib.pbkdf2_hmac('sha256', password.encode(), secure_token.encode(), 100000)
-        
-        if 'pending_registrations' not in request.session:
-            request.session['pending_registrations'] = {}
-        request.session['pending_registrations'][secure_token] = {
-            'email': email,
-            'password_hash': password_hash.hex(),  # Store as hex, not plaintext
-            'created_at': time.time(),
-        }
-        request.session.modified = True
-
-        # Send OTP email
-        try:
-            send_mail(
-                'Your Graphify Verification Code',
-                f'Welcome to Graphify! Your one-time password is: {otp}\n\nThis code expires in 5 minutes.',
-                settings.DEFAULT_FROM_EMAIL,
-                [email],
-                fail_silently=False,
-            )
-            messages.success(request, f"A verification code has been sent to {email}.")
-            return redirect('verify_otp')
-        except Exception as e:
-            logging.error(f"Email send failed (non-sensitive): OTP delivery failed")
-            messages.error(request, "Failed to send verification code. Please try again.")
-            return render(request, 'main/register.html')
-
-    return render(request, 'main/register.html')
-
-@csrf_protect
-def otp_verify_view(request):
-    """Verifies the OTP and creates the user account in Firebase - SECURE VERSION."""
-    reg_data = request.session.get('registration_data')
-    if not reg_data:
-        messages.error(request, "Session expired. Please try registering again.")
-        return redirect('register')
-
-    if request.method == 'POST':
-        submitted_otp = request.POST.get('otp', '')
-        
-        # Check OTP expiry
-        if time.time() > reg_data.get('otp_expiry', 0):
-            messages.error(request, "OTP has expired. Please register again.")
-            _clear_registration_session(request)
-            return redirect('register')
-        
-        # Verify OTP using constant-time comparison (prevents timing attacks)
-        submitted_otp_hash = hashlib.sha256(submitted_otp.encode()).hexdigest()
-        stored_otp_hash = reg_data.get('otp_hash', '')
-        
-        if not secrets.compare_digest(submitted_otp_hash, stored_otp_hash):
-            messages.error(request, "Invalid OTP. Please try again.")
-            return render(request, 'main/verify_otp.html', {'email': reg_data.get('email')})
-        
-        try:
-            # Retrieve password hash from pending registrations
-            pending_regs = request.session.get('pending_registrations', {})
-            secure_token = reg_data.get('secure_token')
-            pending_user = pending_regs.get(secure_token, {})
-            
-            # Verify the pending registration hasn't expired (10 minute max)
-            if time.time() - pending_user.get('created_at', 0) > 600:
-                messages.error(request, "Registration session expired. Please try again.")
-                _clear_registration_session(request)
-                return redirect('register')
-            
-            stored_password_hash = pending_user.get('password_hash')
-            email = reg_data.get('email')
-            
-            if not stored_password_hash or not email:
-                messages.error(request, "Registration data corrupted. Please try again.")
-                _clear_registration_session(request)
-                return redirect('register')
-            
-            # SECURITY: We don't actually recreate the password hash for verification
-            # Instead, we proceed with Firebase Auth creation since OTP was verified
-            # Password will be set directly in Firebase, bypassing session storage entirely
-            
-            # Get the original password from the client-side submit if available
-            # For maximum security, the password should be sent fresh from the form
-            # This view assumes the OTP is sufficient proof of email ownership
-            password = request.POST.get('password', '')
-            password_confirm = request.POST.get('password_confirm', '')
-            
-            if password and password_confirm and password == password_confirm:
-                # Create user in Firebase Auth (this is the source of truth)
-                try:
-                    firebase_user = auth.create_user(
-                        email=email,
-                        password=password
-                    )
-                except auth.EmailAlreadyExistsError:
-                    logging.warning(f"Email registration race condition (non-sensitive)")
-                    messages.error(request, "This email was registered by another session. Please login.")
-                    _clear_registration_session(request)
-                    return redirect('login')
-                
-                # Store additional data in Firestore with Firebase UID as document ID
-                db = firestore.client()
-                user_doc = db.collection('users').document(firebase_user.uid)
-                user_doc.set({
-                    'email': email,
-                    'firebase_uid': firebase_user.uid,
-                    'providers': ['email'],
-                    'created_at': firestore.SERVER_TIMESTAMP
-                })
-                
-                # Create a minimal Django user for session management (use Firebase UID)
-                django_user, created = User.objects.get_or_create(
-                    username=firebase_user.uid,
-                    defaults={
-                        'email': email,
-                        'first_name': '',
-                        'last_name': ''
-                    }
-                )
-                
-                # Set unusable password (Firebase handles authentication)
-                django_user.set_unusable_password()
-                django_user.save()
-                
-                # Clear all registration data securely
-                _clear_registration_session(request)
-                
-                login(request, django_user)
-                return redirect('dashboard')
-            else:
-                messages.error(request, "Passwords must be provided and match.")
-                return render(request, 'main/verify_otp.html', {'email': email})
-                
-        except Exception as e:
-            logging.error(f"Registration verification error (non-sensitive)")
-            messages.error(request, "Registration failed. Please try again.")
-            return redirect('register')
-
-    return render(request, 'main/verify_otp.html', {'email': reg_data.get('email')})
-
-
-def _clear_registration_session(request):
-    """Securely clears all registration-related session data."""
-    reg_data = request.session.get('registration_data', {})
-    secure_token = reg_data.get('secure_token')
-    
-    # Remove from pending_registrations if exists
-    if secure_token and 'pending_registrations' in request.session:
-        if secure_token in request.session['pending_registrations']:
-            del request.session['pending_registrations'][secure_token]
-    
-    # Remove registration_data
-    if 'registration_data' in request.session:
-        del request.session['registration_data']
